% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_default_project.R
\name{initialise_default_project}
\alias{initialise_default_project}
\alias{initialize_default_project}
\title{Initialise a project directory and its config file}
\usage{
initialise_default_project(path = "project")

initialize_default_project(path = "project")
}
\arguments{
\item{path}{an absolute or relative path pointing to the location you would
like the project directory to be created}
}
\description{
Initialise a project directory and its config file
}
