% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_default_project.R
\name{initialize_default_project}
\alias{initialize_default_project}
\title{Initialize a project directory and its config file}
\usage{
initialize_default_project(path = "project")
}
\arguments{
\item{path}{an absolute or relative path pointing to the location you would
like the project directory to be created}
}
\description{
Initialize a project directory and its config file
}
