% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_default_project.R
\name{initialise_default_project}
\alias{initialise_default_project}
\alias{initialize_default_project}
\title{Initialise a project directory and its config file}
\usage{
initialise_default_project(path = "project")

initialize_default_project(path = "project")
}
\arguments{
\item{path}{an absolute or relative path pointing to the location you would
like the project directory to be created}
}
\value{
\code{initialise_default_project()} returns \code{NULL} invisibly. It is called for its
side-effect of creating a default project directory at the specified path.
}
\description{
\code{initialise_default_project()} sets up a default project directory, including
a project directory, a default \code{config.yml} configuration file, an \code{input}
sub-directory, and a \code{loanbooks} sub-directory.
}
\examples{
\dontrun{
project_dir <- "path/to/project"
initialise_default_project(project_dir)
}

}
